Conflicts:
	Equilobe.DailyReport.Models/Equilobe.DailyReport.Models.csproj
	Equilobe.DailyReport.Models/Interfaces/IJiraService.cs
	Equilobe.DailyReport.Models/ReportFrame/JiraReport.cs
	Equilobe.DailyReport.Models/Storage/SourceControlOptions.cs
	Equilobe.DailyReport.SL/Equilobe.DailyReport.SL.csproj
	Equilobe.DailyReport.Utils/Equilobe.DailyReport.Utils.csproj
	JiraReporter/SourceControl/GitHubReportSourceControl.cs
	SvnLogReporter/GitHubReport.cs
	SvnLogReporter/SvnReport.cs

+++ Equilobe.DailyReport.Models/Storage/JiraPolicy.cs
@@ -1,9 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Xml.Serialization;
 
 namespace Equilobe.DailyReport.Models.Storage
@@ -16,7 +13,7 @@ namespace Equilobe.DailyReport.Models.Storage
         public string Password { get; set; }
         public string SharedSecret { get; set; }
 
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         public string DraftEmails { get; set; }
         public string Emails { get; set; }
         public string ReportTime { get; set; }
